Обновление лабораторной #3
Обновление лабораторной, чтобы была возможность запуска с помощью Maven в соостветсвии с быстрым стартом
