+++ lab03/Counter/src/main/java/com/vlsu/lab03/impl/PrimeFactorsCount.java
@@ -1,0 +1,26 @@
+package com.vlsu.lab03.impl;
+
+import com.vlsu.lab03.interfaces.Counter;
+
+public class PrimeFactorsCount implements Counter {
+    public int count(int number) {
+        int result = 0;
+        for(int i = 2; i <= number; i++) {
+            if(number % i == 0) {
+                if(isPrime(i)) {
+                    result++;
+                }
+            }
+        }
+        return result;
+    }
+    
+    public static boolean isPrime(int number) {
+        for (int i = 2; i < number; i++) {
+            if (number % i == 0) {
+                return false;
+            }
+        }
+        return true;
+    }
+}
