+++ lab03/Counter/src/main/java/com/vlsu/lab03/interfaces/Counter.java
@@ -1,0 +1,5 @@
+package com.vlsu.lab03.interfaces;
+
+public interface Counter {
+    int count(int number);  
+}
