+++ lab03/OOP Carshop/src/main/java/com/vlsu/lab03/carshop/cars/Ford.java
@@ -1,0 +1,18 @@
+package com.vlsu.lab03.carshop.cars;
+
+public class Ford extends Car {
+    public int year;
+    public int manufacturerDiscount;
+    
+    public Ford (int speed, double regularPrice, String color, int year, int manufacturerDiscount)
+    {
+        super(speed,regularPrice,color);
+        this.year = year;
+        this.manufacturerDiscount = manufacturerDiscount;
+    }
+
+    public double getSalePrice()
+    {
+        return regularPrice - manufacturerDiscount;
+    }
+}
