+++ lab03/OOP Carshop/src/main/java/com/vlsu/lab03/carshop/cars/Truck.java
@@ -1,0 +1,16 @@
+package com.vlsu.lab03.carshop.cars;
+
+public class Truck extends Car {
+    public int weight;
+    
+    public Truck (int speed, double regularPrice, String color, int weight)
+    {
+        super(speed,regularPrice,color);
+        this.weight = weight;
+    }
+    
+    public double getSalePrice()
+    {
+        return (weight > 2000) ? regularPrice * 0.9 : regularPrice;
+    }
+}
