+++ lab03/OOP Carshop/src/main/java/com/vlsu/lab03/carshop/interfaces/Admin.java
@@ -1,0 +1,5 @@
+package com.vlsu.lab03.carshop.interfaces;
+
+public interface Admin {
+    double getIncome();
+}
