+++ lab03/OOP Carshop/src/main/java/com/vlsu/lab03/carshop/interfaces/Customer.java
@@ -1,0 +1,9 @@
+package com.vlsu.lab03.carshop.interfaces;
+
+public interface Customer {
+    double[] getCarsPrice();
+    String[] getCarsColor();
+    double getCarPrice(int id);
+    String getCarColor(int id);
+    boolean purchaseCar(int id);
+}
