--- lab03/source/src/main/java/carshop/interfaces/Admin.java
@@ -1,10 +1,0 @@
-/*
- * To change this license header, choose License Headers in Project Properties.
- * To change this template file, choose Tools | Templates
- * and open the template in the editor.
- */
-package carshop.interfaces;
-
-public interface Admin {
-    double getIncome();
-}
