--- lab03/source/src/main/java/carshop/interfaces/Customer.java
@@ -1,14 +1,0 @@
-/*
- * To change this license header, choose License Headers in Project Properties.
- * To change this template file, choose Tools | Templates
- * and open the template in the editor.
- */
-package carshop.interfaces;
-
-public interface Customer {
-    double[] getCarsPrice();
-    String[] getCarsColor();
-    double getCarPrice(int id);
-    String getCarColor(int id);
-    boolean purchaseCar(int id);
-}
